Add and remove icon with settings
