+++ Text-Grab/Views/SettingsWindow.xaml
@@ -115,6 +115,17 @@
                         Never automatically add text to the clipboard
                     </TextBlock>
                 </CheckBox>
+
+                <CheckBox
+                    Name="RunInBackgroundChkBx"
+                    Margin="0,16,0,0"
+                    VerticalContentAlignment="Center"
+                    Style="{StaticResource labelText}">
+                    <TextBlock Style="{StaticResource TextBodyNormal}">
+                        Run Text Grab in the background
+                    </TextBlock>
+                </CheckBox>
+
             </StackPanel>
             <StackPanel Grid.Row="1" Orientation="Horizontal">
                 <controls:CollapsibleButton
